[fix] 유저 (#17)
User 의 id 필드 Integer -> Long으로 타입 변경
=> Controller, Service, Repository, Response 클래스 id 타입 수정

+++ src/main/java/co/kr/metacoding/backendtest/user/UserController.java
@@ -16,13 +16,13 @@ public class UserController {
     }
 
     @GetMapping("/users/{id}")
-    public ResponseEntity<?> getDetail(@PathVariable("id") Integer id) {
+    public ResponseEntity<?> getDetail(@PathVariable("id") Long id) {
         UserResponse.DTO respDTO = userService.getDetail(id);
         return ResponseEntity.ok(respDTO);
     }
 
     @PutMapping("/users/{id}")
-    public ResponseEntity<?> update(@PathVariable("id") Integer id, @RequestBody UserRequest.UpdateDTO reqDTO){
+    public ResponseEntity<?> update(@PathVariable("id") Long id, @RequestBody UserRequest.UpdateDTO reqDTO){
         UserResponse.DTO respDTO = userService.update(id, reqDTO);
         return ResponseEntity.ok(respDTO);
     }
