+++ src/main/java/co/kr/metacoding/backendtest/user/UserService.java
@@ -25,14 +25,14 @@ public class UserService {
         return new UserResponse.SaveDTO(userPS);
     }
 
-    public UserResponse.DTO getDetail(Integer id) {
+    public UserResponse.DTO getDetail(Long id) {
         User userOP = userRepository.findById(id)
                 .orElseThrow(() -> new ExceptionApi404("해당 유저를 찾을 수 없습니다."));
         return new UserResponse.DTO(userOP);
     }
 
     @Transactional
-    public UserResponse.DTO update(Integer id, UserRequest.UpdateDTO reqDTO) {
+    public UserResponse.DTO update(Long id, UserRequest.UpdateDTO reqDTO) {
         // 유저 존재 여부 확인
         User userPS = userRepository.findById(id)
                 .orElseThrow(() -> new ExceptionApi404("해당 유저를 찾을 수 없습니다."));
